+++ packages/core/js/breakpoints.ts
@@ -1,0 +1,7 @@
+export default {
+  xSmall: '640px',
+  small: '768px',
+  medium: '1024px',
+  large: '1280px',
+  xLarge: '1600px'
+}