+++ packages/util/src/combine-fns.js
@@ -1,7 +1,5 @@
+import { isFunction } from '.'
+
 export function combineFns(...fns) {
   return (...args) => fns.filter(isFunction).forEach(fn => fn(...args))
-}
-
-function isFunction(fn) {
-  return typeof fn === 'function'
 }